start editor
