+++ js/main.js
@@ -1,6 +1,6 @@
 import {pictures} from './data.js';
 import {renderPictures} from './picture.js';
-import  './editImage.js';
+import  './openEditor.js';
 renderPictures(pictures);
 // console.log(pictures.find((item) => item.id === 1));
